탈퇴 유효성검사

+++ src/main/java/com/m2j2/haruseoul/member/service/DefaultMemberService.java
@@ -47,8 +47,10 @@ public class DefaultMemberService implements MemberService {
 
     @Override
     public void delete(Long id) {
+        if (!memberRepository.existsById(id)) {
+            throw new IllegalArgumentException("해당 ID가 존재하지 않습니다.이미 삭제되었거나 없는 계정입니다.");
+        }
         memberRepository.deleteById(id);
-
     }
 
 }
